Bufferevent MUST NOT die when used by C code
Problem: I experimentally noticed, when testing mkok-onion-poller, that
calling bufferevent_free() during a callback over a SSL buffer is triggering
various use-after-free errors in the bufferevent SSL code.

Diagnosis: I claimed to guarantee that, whenever a piece of code is used,
it cannot be deleted, because I use Var<T>. However, the C code uses a raw
pointer, not a smart pointer. So the previous statement is false.

Fix:

1. Pass a `Var<T> *` pointer to C code, such that the C code keeps a
   reference on our data structures and we're safe.

   This is implemented in socket_new() and openssl_filter_new().

2. Detect in C code when the `Var<T>` is no longer used by C++ because
   its usage count is now 1 (i.e. only C used that struct).

   This is implemented in all the three bufferevent C callbacks by calling
   a new function named mkok_libevent_bev_is_ignored_by_cxx().

3. In such case delete the `Var<T> *` pointer.

   This is the job of mkok_libevent_is_ignored_by_cxx().

The most complicated step is the latter:

a. It turns out, when we're in 3, we still cannot safely call delete because
   still there is underlying code not liking that. Hence, we call `once()`
   to dispose of the object in the following I/O loop cycle.

   Before disposing of the object, anyway, we put it into a state that cannot
   hurt, i.e., no callbacks, all I/O events disabled.

   Both measures are implemented in mkok_libevent_is_ignored_by_cxx(). We also
   added an `evbase` field to `Bufferevent` to call `once()`.

b. It still turns out that now we cannot chain the destruction of SSL and of
   underlying as it used to be. If we store a reference to the underlying into
   the SSL structure, in fact, the usage count of the underlying is not going
   to become 1 when we leave C++ code. So, devise another strategy by which:

       - we steal the `bevp` of underlying in openssl_filter_new()

       - we use the destructor of Bufferevent to destroy and we deal with
         the case where `bevp` is nullptr

       - we adapt a. for the case where `bevp` is nullptr

c. Since we need an `evbase` field in a., entiretly dispose of the previous
   strategy in which destruction was performed inside the lambda, one of the
   reasons of which was to keep alive `evbase`.

This should be it. Hopefully.